Move out some utility functions.
