+++ Wrappers/Python/cil/framework/__init__.py
@@ -23,7 +23,7 @@ from .framework import ImageData
 from .AcquisitionData import AcquisitionData
 from .framework import ImageGeometry, AcquisitionGeometry
 from .Vectors import VectorData, VectorGeometry
-from .framework import find_key, message
+from .utils import find_key, message
 from .processors import Processor, DataProcessor, AX, CastDataContainer, PixelByPixelDataProcessor
 from .BlockDataContainer import BlockDataContainer
 from .BlockGeometry import BlockGeometry
